+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Health-check API endpoint `/api/health` with SMTP provider check
+
 ## [0.11.1] - 2026-01-19
 Minor release with security dependencies updates and some internal refactoring.